Re-add support for older ruby and rails versions
Ruby 2.6 and Rails 5.2 are no longer supported and have reached their
end of life, but someone may still use them and they are still
compatible with this gem and as of now it doesn't cost us anything to
still support them.

And I do not want to make a major release so often, just because a new
version of ruby or rails has been released (assuming there are no
breaking changes for Frikandel to support the new versions).